add base class
