+++ snuba/datasets/storages/transactions.py
@@ -25,7 +25,9 @@ from snuba.query.processors.mapping_optimizer import MappingOptimizer
 from snuba.query.processors.mapping_promoter import MappingColumnPromoter
 from snuba.query.processors.prewhere import PrewhereProcessor
 from snuba.query.processors.uuid_column_processor import UUIDColumnProcessor
-from snuba.query.processors.hexint_column_processor import HexIntColumnProcessor
+from snuba.query.processors.type_converters.hexint_column_processor import (
+    HexIntColumnProcessor,
+)
 from snuba.web.split import TimeSplitQueryStrategy
 
 columns = ColumnSet(
