+++ snuba/query/processors/type_converters/hexint_column_processor.py
@@ -1,0 +1,22 @@
+from snuba.query.expressions import Column, Expression, FunctionCall, Literal
+from snuba.query.processors.type_converters import BaseTypeConverter, ColumnTypeError
+
+
+class HexIntColumnProcessor(BaseTypeConverter):
+    def translate_literal(self, exp: Expression) -> Expression:
+        try:
+            assert isinstance(exp, Literal)
+            assert isinstance(exp.value, str)
+            return Literal(alias=exp.alias, value=int(exp.value, 16))
+        except AssertionError:
+            raise ColumnTypeError("Invalid hexint")
+
+    def process_expressions(self, exp: Expression) -> Expression:
+        if isinstance(exp, Column) and exp.column_name in self.columns:
+            return FunctionCall(
+                exp.alias,
+                "lower",
+                (FunctionCall(None, "hex", (Column(None, None, exp.column_name),),),),
+            )
+
+        return exp
